pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["andrewabrahamian <andrew.abrahamian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
joblib = "^1.3.2"
scikit-learn = "1.3.0"
fastapi-cache2 = {version = "0.1.9", extras = ["redis"]}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
httpx = "^0.24.1"
ruff = "^0.0.287"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
